backend/requirements.txt
# requirements.txt

fastapi==0.101.1
uvicorn[standard]==0.38.0
asyncpg==0.30.0
pydantic==2.11.0
pydantic-settings==2.11.0
python-dotenv==1.1.1
google-generativeai==0.8.5
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
typing-extensions===4.15.0
pytest==8.4.2
pytest-asyncio==1.2.0
httpx==0.28.1
PyJWT==2.10.1
solders==0.27.0